refactor(core): :recycle: move rational conversion logic to conversion module and create core module to break circular dependency
