--- src/sexagesimal_calculator/contants.py
@@ -1,4 +1,0 @@
-# Constants
-BASE = 60
-PART_SEP = ";"
-VAL_SEP = ","
